requirements.txt
torch==2.10.0
transformers==5.1.0
safetensors==0.7.0
grpcio==1.78.0
grpcio-tools==1.78.0
protobuf==6.33.5
fastapi==0.129.0
uvicorn==0.40.0
python-multipart==0.0.22
websockets==15.0.1
web3==7.14.1
cryptography==46.0.5
numpy==2.4.2
requests==2.32.5
pytest==9.0.2
